refactor(client): extract websocket send pipeline (#1118)
Moves the multi-device DM encryption fanout, the fallback chain, the
device-contents builder, the wire-frame builders (DM + group), the
group key self-heal, GRP1/GRP2 dispatch, the failed-message helper,
and the friendly-error mapper into a part file under
lib/src/providers/websocket/. The host class keeps only the public
sendMessage / sendGroupMessage entry points (thin delegations), so the
Notifier surface area is unchanged. Wire-shape contract test plus
~70 send-path tests stay green.
